+++ config/application.rb
@@ -27,6 +27,6 @@ module Zssn
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
 
-    config.version = '1.1.0'
+    config.version = '1.2.0'
   end
 end